requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
Deprecated==1.2.13
distro==1.8.0
Django==4.1.7
django-cleanup==7.0.0
django-extensions==3.2.1
django-filter==22.1
django-image-uploader-widget==0.2.3
django-redis==5.2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
fake-useragent==1.1.1
gunicorn==20.1.0
idna==3.4
Markdown==3.4.1
netaddr==0.8.0
Pillow==9.4.0
psycopg2==2.9.5
python-dotenv==0.21.1
pytz==2022.7.1
redis==4.5.2
requests==2.28.2
soupsieve==2.4
sqlparse==0.4.3
urllib3==1.26.14
wrapt==1.15.0
yookassa==2.3.5
